symbolic_fuzzing/requirements.txt
# Requirements for DifuzzRTL Symbolic Fuzzing System
# 
# Core symbolic execution engine
# Note: angr and its dependencies can be large and may require compilation
# If installation fails, try: pip install --no-cache-dir angr

# Core angr binary analysis framework
angr>=9.2.0,<10.0.0

# Binary analysis and manipulation components
capstone>=4.0.2,<6.0.0
archinfo>=9.2.0,<10.0.0
pyvex>=9.2.0,<10.0.0
cle>=9.2.0,<10.0.0
claripy>=9.2.0,<10.0.0

# Constraint solving
z3-solver>=4.8.0,<5.0.0

# System and process management
psutil>=5.8.0

# Networking and data handling (used by angr)
requests>=2.25.0
networkx>=2.5,<4.0

# Binary file handling
pyelftools>=0.27

# Configuration and serialization
pyyaml>=5.4.0

# Colored output for better UX
colorama>=0.4.4

# Development and testing dependencies
pytest>=6.0.0
pytest-cov>=2.10.0

# Optional performance improvements
# Uncomment these if you want better performance or encounter issues:
# 
# unicorn>=1.0.2  # Unicorn engine for faster emulation
# keystone-engine>=0.9.2  # Keystone assembler engine
# 
# Optional visualization and analysis tools:
# matplotlib>=3.3.0  # For plotting coverage graphs
# numpy>=1.20.0      # For numerical analysis
# pandas>=1.3.0      # For data analysis

# Platform-specific notes:
# - On Ubuntu/Debian, you may need: sudo apt-get install build-essential libffi-dev
# - On CentOS/RHEL, you may need: sudo yum groupinstall "Development Tools" && sudo yum install libffi-devel
# - On macOS, you may need: xcode-select --install
